+++ bella-language-server/server/src/handlers/document-symbol.handler.ts
@@ -1,13 +1,24 @@
-import * as LSP from 'vscode-languageserver'
+import {DocumentSymbolParams, WorkspaceSymbolParams, SymbolInformation} from 'vscode-languageserver'
 import { LSPDeclarationRegistry } from '../lsp-declaration-registry';
 
 export class DocumentSymbolHandler {
     constructor(private cache: LSPDeclarationRegistry) {
         this.cache = cache;
     }
-    public findSymbols(params: LSP.DocumentSymbolParams): LSP.SymbolInformation[] {
+    public findSymbols(params: DocumentSymbolParams): SymbolInformation[] {
         let docUri = params.textDocument.uri;
-        return this.cache.getDeclarations(docUri);;
-        // return this.analyzer.findSymbols(params.textDocument.uri);
+        return this.cache.getLSPDeclarations(docUri);
+    }
+
+    public findSymbolsInWorkspace(params: WorkspaceSymbolParams): SymbolInformation[] {
+        return this.cache.getLSPDeclarationsForQuery({
+            uriFilter: {
+                active: false
+            },
+            parentFilter: {
+                active: true,
+                hasParent: false
+            }
+        });
     }
 }