+++ bella-language-server/server/webpack.config.js
@@ -17,7 +17,7 @@ const config = {
         devtoolModuleFilenameTemplate: '[absolute-resource-path]'
     },
     // devtool: 'eval-source-map',
-    devtool: 'inline-source-map',
+    devtool: 'source-map',
     // externals: {
     //     vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
     // },